app/wiki/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
psycopg2-binary
alembic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
httpx
jinja2
python-slugify
